ENH: adds authorization to all pages
