+++ frontend/src/App.tsx
@@ -1,8 +1,25 @@
-import { Outlet } from "react-router-dom";
+import { useEffect, useState } from "react";
+import { Outlet, useNavigate } from "react-router-dom";
 import "./App.css";
 import Footer from "./components/utilities/Footer";
 
 function App() {
+  const navigator = useNavigate();
+  const [page, setPage] = useState([]);
+  const fetchPage = async () => {
+    const response = await fetch("https://backend-mu-plum.vercel.app/", {
+      headers: {
+        Authorization: `Bearer ${authToken}`,
+      },
+    });
+    if (!response.ok) {
+      navigator("/auth/signin");
+    }
+  };
+  const authToken = localStorage.getItem("authToken");
+  useEffect(() => {
+    fetchPage();
+  }, page);
   return (
     <div className="App" data-testid="App">
       <Outlet></Outlet>
